Renamed applyCallback to simplifyCallback
